refactor(core): combine three core files to one, rename core/plugin module to plugin
